Project initialization
This is a project refresh for EATSv5 which will create an audit trail of a fresh WordPress plugin from a standard WP plugin base. It provides a clean basis for comparison withother WP plugin projects in terms of software related code capability, cost, complexity, durability, and related meterics.

The 2008 WP PHP EATS is then being overlaid on the current WP5+/PHP7+ operating environment.